Exception classes structure refactored
